Cleanup and document.

+++ string.h
@@ -59,8 +59,8 @@ inline const string& str(const string& s) { return s; }
 /// Converts integers
 template<uint base=10> string utoa(uint64 number, int pad=0);
 template<uint base=10> string itoa(int64 number, int pad=0);
-inline string bin(uint n, int pad=0) { return utoa<2>(n,pad); }
-inline string dec(int n, int pad=0) { return itoa<10>(n,pad); }
+inline string bin(uint64 n, int pad=0) { return utoa<2>(n,pad); }
+inline string dec(int64 n, int pad=0) { return itoa<10>(n,pad); }
 inline string str(const uint8& n) { return dec(n); }
 inline string str(const int8& n) { return dec(n); }
 inline string str(const uint16& n) { return dec(n); }
